pyproject.toml
[project]
name = "nonebot_plugin_bililivedm"
version = "0.1.1"
description = "基于ws的b站弹幕监控"
authors = [
  { name="STESmly", email="STESmly@mail.com" },
]
readme = "README.md"
requires-python = ">=3.8, <4.0"
dependencies = [
    "nonebot2>=2.2.1",
    "aiohttp>=3.10.10",
    "Brotli>=1.1.0",
    "yarl>=1.15.2",
]

[project.urls]
Homepage = "https://github.com/STESmly/nonebot_plugin_bililivedm"
Repository = "https://github.com/STESmly/nonebot_plugin_bililivedm"

[tool.pdm]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"